+++ packages/react-native-client/src/common/client.ts
@@ -1,5 +1,5 @@
-import { Metadata } from '../RNFishjamClient.types';
-import RNFishjamClientModule from '../RNFishjamClientModule';
+import { Metadata } from '../types';
+import RNFishjamClientModule from '../FishjamClient';
 
 export const connect = (
   url: string,
