+++ packages/react-native-client/src/components/VideoRendererView.tsx
@@ -2,7 +2,7 @@ import { requireNativeViewManager } from 'expo-modules-core';
 import * as React from 'react';
 import { ViewStyle } from 'react-native';
 
-import { VideoLayout } from '../RNFishjamClient.types';
+import { VideoLayout } from '../types';
 
 export type VideoRendererProps = {
   /**
