+++ packages/react-native-client/src/hooks/useCamera.ts
@@ -7,8 +7,8 @@ import {
   SimulcastConfig,
   TrackBandwidthLimit,
   TrackEncoding,
-} from '../RNFishjamClient.types';
-import RNFishjamClientModule from '../RNFishjamClientModule';
+} from '../types';
+import RNFishjamClientModule from '../FishjamClient';
 import { ReceivableEvents, eventEmitter } from '../common/eventEmitter';
 
 type IsCameraOnEvent = { IsCameraOn: boolean };
