+++ packages/react-native-client/src/hooks/useMicrophone.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 
-import { Metadata } from '../RNFishjamClient.types';
-import RNFishjamClientModule from '../RNFishjamClientModule';
+import { Metadata } from '../types';
+import RNFishjamClientModule from '../FishjamClient';
 import { ReceivableEvents, eventEmitter } from '../common/eventEmitter';
 
 export type MicrophoneConfig<MetadataType extends Metadata> = {
