+++ packages/react-native-client/src/index.tsx
@@ -58,4 +58,4 @@ export { VideoPreviewView } from './components/VideoPreviewView';
 export type { VideoRendererProps } from './components/VideoRendererView';
 export { VideoRendererView } from './components/VideoRendererView';
 
-export * from './RNFishjamClient.types';
+export * from './types';
